Refactor App.jsx, utils and APIs
