+++ .gitignore
@@ -10,3 +10,4 @@ classpath
 .DS_Store            
 node_modules
 dist
+npm-debug.log
